# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "masterwork",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/proposals/{proposal_id}/accept": {
            "post": {
                "tags": [
                    "proposals"
                ],
                "summary": "Accept Proposal",
                "operationId": "acceptProposal",
                "parameters": [
                    {
                        "name": "proposal_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Proposal Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposal"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/proposals/{proposal_id}/reject": {
            "post": {
                "tags": [
                    "proposals"
                ],
                "summary": "Reject Proposal",
                "operationId": "rejectProposal",
                "parameters": [
                    {
                        "name": "proposal_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Proposal Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposal"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/simulations": {
            "get": {
                "tags": [
                    "simulations"
                ],
                "summary": "List Simulations",
                "operationId": "listSimulations",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Simulation"
                                    },
                                    "title": "Response Listsimulations"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "simulations"
                ],
                "summary": "Create Simulation",
                "operationId": "createSimulation",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SimulationCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Simulation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/simulations/autopilot": {
            "post": {
                "tags": [
                    "simulations"
                ],
                "summary": "Start Autopilot",
                "operationId": "startSimulationAutopilot",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AutopilotCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Simulation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/simulations/autopilot/{run_id}/stop": {
            "post": {
                "tags": [
                    "simulations"
                ],
                "summary": "Stop Autopilot",
                "operationId": "stopSimulationAutopilot",
                "parameters": [
                    {
                        "name": "run_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Run Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/simulations/scenario": {
            "post": {
                "tags": [
                    "simulations"
                ],
                "summary": "Generate Scenario",
                "operationId": "generateSimulationScenario",
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Project Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScenarioGenerateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/simulations/{simulation_id}": {
            "get": {
                "tags": [
                    "simulations"
                ],
                "summary": "Get Simulation",
                "operationId": "getSimulation",
                "parameters": [
                    {
                        "name": "simulation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Simulation Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Simulation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "simulations"
                ],
                "summary": "Delete Simulation",
                "operationId": "deleteSimulation",
                "parameters": [
                    {
                        "name": "simulation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Simulation Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/simulations/{simulation_id}/suggestions/{suggestion_index}/apply": {
            "post": {
                "tags": [
                    "simulations"
                ],
                "summary": "Apply Suggestion",
                "operationId": "applySimulationSuggestion",
                "parameters": [
                    {
                        "name": "simulation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Simulation Id"
                        }
                    },
                    {
                        "name": "suggestion_index",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Suggestion Index"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Simulation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AutopilotCreateRequest": {
                "properties": {
                    "scenario": {
                        "type": "string",
                        "title": "Scenario",
                        "description": "Optional scenario to simulate; empty lets the model derive one from the goal.",
                        "default": ""
                    },
                    "control_run": {
                        "type": "boolean",
                        "title": "Control Run",
                        "description": "Build the capability checklist from scratch instead of re-grading the previous run's. Forced anyway when the last run of this scenario scored 100. Applies to the first iteration only.",
                        "default": false
                    },
                    "iterations": {
                        "type": "integer",
                        "maximum": 20.0,
                        "minimum": 1.0,
                        "title": "Iterations",
                        "description": "Maximum number of chained runs; stops early when a run has no suggestions.",
                        "default": 5
                    }
                },
                "type": "object",
                "title": "AutopilotCreateRequest"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Proposal": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ProposalStatus"
                    },
                    "summary": {
                        "type": "string",
                        "title": "Summary"
                    },
                    "changes": {
                        "items": {
                            "$ref": "#/components/schemas/ProposalChange"
                        },
                        "type": "array",
                        "title": "Changes"
                    },
                    "project_update": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ProjectUpdate"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "A proposed project update; carried alongside or instead of file changes."
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error",
                        "description": "Set when status is 'failed'."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "status",
                    "summary",
                    "changes",
                    "created_at"
                ],
                "title": "Proposal"
            },
...
            "ProposalStatus": {
                "type": "string",
                "enum": [
                    "pending",
                    "applied",
                    "rejected",
                    "failed"
                ],
                "title": "ProposalStatus"
            },
            "ScenarioGenerateResponse": {
                "properties": {
                    "scenario": {
                        "type": "string",
                        "title": "Scenario",
                        "description": "The generated scenario, saved on the project."
                    }
                },
                "type": "object",
                "required": [
                    "scenario"
                ],
                "title": "ScenarioGenerateResponse"
            },
            "Simulation": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "project_id": {
                        "type": "string",
                        "title": "Project Id"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "running",
                            "completed",
                            "failed"
                        ],
                        "title": "Status"
                    },
                    "scenario": {
                        "type": "string",
                        "title": "Scenario",
                        "description": "User-provided scenario; empty = derived from goal."
                    },
                    "score": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Score",
                        "description": "0-100, null until completed."
                    },
                    "verdict": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Verdict"
                    },
                    "summary": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Summary",
                        "description": "Markdown: what the simulated run did."
                    },
                    "analysis": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Analysis",
                        "description": "Markdown: strengths, gaps, failure points."
                    },
                    "trace_mermaid": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Trace Mermaid",
                        "description": "Mermaid trace of the simulated run."
                    },
                    "checklist": {
                        "items": {
                            "$ref": "#/components/schemas/SimulationChecklistItem"
                        },
                        "type": "array",
                        "title": "Checklist",
                        "description": "Capability checklist the score is computed from; [] for pre-v1.7 runs."
                    },
                    "suggestions": {
                        "items": {
                            "$ref": "#/components/schemas/SimulationSuggestion"
                        },
                        "type": "array",
                        "title": "Suggestions"
                    },
                    "control_run": {
                        "type": "boolean",
                        "title": "Control Run",
                        "description": "True when this run built its checklist from scratch instead of re-grading the previous run's; forced automatically after a run scores 100."
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "completed_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Completed At"
                    },
                    "autopilot_run_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Autopilot Run Id",
                        "description": "Set when this run is one iteration of an autopilot chain."
                    },
                    "autopilot_iteration": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Autopilot Iteration",
                        "description": "1-based iteration number."
                    },
                    "autopilot_total": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Autopilot Total",
                        "description": "Requested iteration count."
                    },
                    "stats": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/SimulationStats"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "CLI-reported run metadata; null for runs before v1.6."
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "project_id",
                    "status",
                    "scenario",
                    "score",
                    "verdict",
                    "summary",
                    "analysis",
                    "trace_mermaid",
                    "checklist",
                    "suggestions",
                    "control_run",
                    "error",
                    "created_at",
                    "completed_at",
                    "autopilot_run_id",
                    "autopilot_iteration",
                    "autopilot_total",
                    "stats"
                ],
                "title": "Simulation"
            },
...
            "SimulationCreateRequest": {
                "properties": {
                    "scenario": {
                        "type": "string",
                        "title": "Scenario",
                        "description": "Optional scenario to simulate; empty lets the model derive one from the goal.",
                        "default": ""
                    },
                    "control_run": {
                        "type": "boolean",
                        "title": "Control Run",
                        "description": "Build the capability checklist from scratch instead of re-grading the previous run's. Forced anyway when the last run of this scenario scored 100.",
                        "default": false
                    }
                },
                "type": "object",
                "title": "SimulationCreateRequest"
            },
...
        }
    }
}